requirement.txt
django-address
django-background-tasks